requirements.txt
asgiref==3.2.10
Django==1.11.29
django-filter==1.1.0
django-mysql==3.8.1
djangorestframework==3.11.1
djangorestframework-filters==0.11.1
mysqlclient==2.0.1
pkg-resources==0.0.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
